requirements.txt
# Core Dependencies - Updated for Python 3.13 compatibility
fastapi==0.115.6
uvicorn[standard]==0.34.0
motor==3.6.0
pydantic==2.10.6
pydantic-settings==2.7.1
python-dotenv==1.0.1
httpx==0.28.1

# Testing
pytest==8.3.4
pytest-asyncio==0.25.2
pytest-cov==6.0.0

# Code Quality
mypy==1.14.1
ruff==0.9.3

# AI Integration
openai==1.59.8
anthropic==0.42.0

# Utilities
python-multipart==0.0.20
